Move the test classes doing only setup in common.py
Otherwise, testsuites that use them have to import the test_synchronization
testsuite and that run the tests even if they are not asked for.
